+++ testauth/settings_core.py
@@ -257,7 +257,6 @@ DEBUG = False
 # Add any additional apps to this list.
 INSTALLED_APPS += [
     "structures",
-    "eveuniverse",
 ]
 
 # Enter credentials to use MySQL/MariaDB. Comment out to use sqlite3